Added `de/enums` module
