+++ crates/bevy_reflect/src/serde/de/deserializer.rs
@@ -1,11 +1,12 @@
 use crate::serde::de::arrays::ArrayVisitor;
+use crate::serde::de::enums::EnumVisitor;
 use crate::serde::de::lists::ListVisitor;
 use crate::serde::de::maps::MapVisitor;
 use crate::serde::de::sets::SetVisitor;
 use crate::serde::de::structs::StructVisitor;
 use crate::serde::de::tuple_structs::TupleStructVisitor;
 use crate::serde::de::tuples::TupleVisitor;
-use crate::serde::de::{EnumVisitor, OptionVisitor};
+use crate::serde::de::OptionVisitor;
 use crate::serde::TypeRegistrationDeserializer;
 use crate::{PartialReflect, ReflectDeserialize, TypeInfo, TypeRegistration, TypeRegistry};
 use core::fmt::Formatter;
@@ -318,11 +319,7 @@ impl<'a, 'de> DeserializeSeed<'de> for TypedReflectDeserializer<'a> {
                     deserializer.deserialize_enum(
                         enum_info.type_path_table().ident().unwrap(),
                         enum_info.variant_names(),
-                        EnumVisitor {
-                            enum_info,
-                            registration: self.registration,
-                            registry: self.registry,
-                        },
+                        EnumVisitor::new(enum_info, self.registration, self.registry),
                     )?
                 };
                 dynamic_enum.set_represented_type(Some(self.registration.type_info()));
